pyproject.toml
[tool.poetry]
name = "audprodassetdb"
version = "0.1.0"
description = "A tool to build an indexed database of your assets to use in audio production."
authors = ["Daniel Andrlik <daniel@andrlik.org>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = "^0.19.1"
aerich = "^0.6.3"
uvloop = "^0.16.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
safety = "^1.10.3"
darglint = "^1.8.1"
bandit = "^1.7.4"
black = {extras = ["d"], version = "^22.3.0"}
mkdocs = "^1.3.0"
mkdocs-material = "^8.2.16"
mkdocs-include-markdown-plugin = "^3.5.2"
mkdocstrings = "^0.19.0"
pytest = "^7.1.2"
pytest-sugar = "^0.9.4"
pytest-cov = "^3.0.0"
colorama = "^0.4.4"
pylint = "^2.14.0"
python-language-server = {extras = ["all"], version = "^0.36.2"}


[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | migrations
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
testpaths = ["tests", "audprodassetdb"]
python_files = ["tests.py", "test_*.py"]
norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__", ".venv"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
  "--cov=audprodassetdb",
  "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["audprodassetdb"]
omit = ["*migrations*", "*tests*"]

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.9
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true
skip_glob = "**/migrations/*.py"

[tool.pydocstyle]
ignore = "D100"
